Working Search bar
The search bar is rudimentary, but working, I do plan on changing it later, but you can search things up, if we really wanted we could get rid of the search page, and implement this into the index page. This does need a large amount of work, I  did add a get method to the search bar to use. the search algorithm  is all currently sitting inside the searcher.js's ajax function. I had to make it work ,  later we can make it better.